for discord, font changer!
Semi broken. If you type too fast it might mess up (such as changing order of the text). Also, before sending a message, press space (or else the last letter of the message will be moved to the end). You currently cant delete the letters, or make newlines using the shift+enter method. I definitely will try to improve it in the future. 
I know it doesnt work in dms, i'll fix it later :P